# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "A simple in-memory CRUD API for managing tasks, with Supabase-backed authentication"
  },
  "paths": {
    "/tasks": {
      "get": {
        "summary": "List all tasks",
        "responses": {
          "200": { "description": "A list of tasks" }
        }
      },
      "post": {
        "summary": "Create a new task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" }
                },
                "required": ["title"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Task created" },
          "400": { "description": "Missing or invalid title" }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Get a single task by id",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "The task" },
          "404": { "description": "Task not found" }
        }
      },
      "put": {
        "summary": "Update a task by id",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "done": { "type": "boolean" }
                },
                "required": ["title"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Task updated" },
          "400": { "description": "Missing or invalid title" },
          "404": { "description": "Task not found" }
        }
      },
      "delete": {
        "summary": "Delete a task by id",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "Task deleted" },
          "404": { "description": "Task not found" }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "summary": "Create a new user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Missing fields or user already registered" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Authenticate and return a JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful, returns access and refresh tokens" },
          "400": { "description": "Missing email or password" },
          "401": { "description": "Invalid login credentials" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "End the user's session",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Logged out successfully" },
          "401": { "description": "Missing or invalid token" }
        }
      }
    },
    "/public/info": {
      "get": {
        "summary": "Public, open data",
        "responses": {
          "200": { "description": "Public info message" }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Read private profile data",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "The authenticated user's profile" },
          "401": { "description": "Missing or invalid token" }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "summary": "Protected dashboard route (demonstrates middleware reuse)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Dashboard welcome message" },
          "401": { "description": "Missing or invalid token" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}